backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai>=1.0.0
requests==2.31.0
Pillow==10.0.0
numpy==1.24.0
pandas==2.0.0
httpx>=0.24.0
